pyproject.toml
[project]
name = "home-assistant"
version = "0.1.0"
description = ""
authors = [
    {name = "Xander te Winkel",email = "572004@student.saxion.nl"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "psutil (>=7.1.3,<8.0.0)",
    "pysmb (>=1.2.13,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "openai (>=2.9.0,<3.0.0)",
    "flask-cors (>=6.0.1,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
